Reformatting metadata
